10715 : Add configurable templates for issue tracker emails 10714 : Render URLs as hyperlinks in grid view 10713 : Add additional admin-configurable string columns in issue tracker
Also, improvements to EmailTemplateService:

Request templates based on java.lang.Class instead of class name. Use generics to avoid casting to subclass.

Allow scoping templates to folders. Currently issue emails are the only ones that make use of this, but it should now be easy to add project-scoped templates for new user emails, etc.

Perform the audit logging as part of MailHelper.send() instead of as a separate call to make it impossible to forget to add logging.

SVN r14286 |2010-08-05 15:58:16 +0000