Migrate assay results from OntologyManager to hard tables for TSV/GPAT, Microarray, CBC, and file-based assays. All hard tables have their own RowIds instead of the old ObjectIds defined by exp.Object. This requires fixing up the values on the dataset side so that copy-to-study links are correctly maintained.
This is done out of the normal upgrade process because AssayProviders need to be registered first. A startup listener is added at upgrade time and performed after the modules have had startup() invoked.

Add support for real multi-column FKs with LookupForeignKey and LookupColumn. They are defined much like the regular, single-column FKs, with a ColumnInfo from the source table and a column name in the target table. This replaces the joinOnContainer setting, which turned out to always join based on the Container column in the query's base table, not necessarily the one in the source of the lookup.

Fill in some gaps in StorageProvisioner, like the ability to create tables with auto-incrementing RowId PKs.

Greatly simplify the joins from assay results to copied-to-study datasets based on the knowledge that there is at most one dataset per assay definition in a target study. Also, don't do the query on the UNION-backed study.StudyData query, do it directly against the correct dataset.

Make DomainKinds Priority based so that they can have weak and strong preferences for recognizing domain URIs.

SVN r15312 |2010-12-28 17:45:43 +0000