Get rid of all but the assay data RowId column in linked datasets as part of an upgrade script.
Simply copy-to-study code significantly. Consolidate to a standard implementation, with the exception of Luminex, which uses row-level LSIDs as SourceLSIDs instead of just the source run.

Add auto-copy-to-study feature. In an assay design, you can optionally choose a study to which all data with ptid/visit values will be linked. Supported for all assay types that support copy-to-study.

Improve copy-to-study perf for large operations by doing a single query to get all the run LSIDs we need, instead of firing off per-row level queries.

Improve compatibility with old column names in datasets.

SVN r15503 |2011-01-30 00:23:39 +0000