Fix ClientAPITest -- properly escape JavaScript strings in .js files used in test.
Details: New jtidy now treats JavaScript strings such as '</foo>' as errors, which ends up stripping the html body completely. Near-term solution is to escape the slashes (with double backslashes?!?). We may want to revert back to an older build and/or report this as a jtidy bug and/or change the test code to handle valid .js files (not require a double backslash for escaping).

SVN r16099 |2011-04-21 04:42:02 +0000