SpecimenImporter refactor:
- Tighten up file stream opening/closing (replace blind opening/closing and passing around of open streams)
- New SpecimenImportFile abstracts source details of simple vs. advanced specimen loads
- Support binary column types and write input_hash if provided in the source data
- New SpecimenTableType enum to replace magic strings for each table/file: enum encapsulate tsv name, table name, and candidate columns
- Additional option for TabLoader to close passed in Reader on complete
- Try-with-resources, Selector, and generic type inference
- TestSet for key specimen tests

And, most importantly, provide a mechanism to inject a SpecimenImportStrategy, which filters the input, insert, and delete operations. Default implementation does no filtering (strategy is to delete everything and insert everything in the new archive, as we've been doing), but other implementations can be used to enable incremental specimen loads and editable specimen repositories.

SVN r26528 |2013-05-22 00:02:12 +0000