Create ServerNotification Components (#884)
Add AppNotification component.
Add component to NavBar.
Add support for the 'Mark all as read' link.
Changed package component is in.
Changed scope of the finder for ServerNotificationItem.
Misc other code review feedback.
 Modified ServerNotificationMenu to return only one message identified by position in list or text of message.
Updated finder for ServerNotificationItem to be able to search based on message text or position.
Updated comments to be consistent when referring to a notification item vs. the message part of the item.